+++ custom_components/solarcharger/chargers/charger_chargeable_base.py
@@ -9,23 +9,23 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntry
 
 from ..const import (
-    ENTITY_CHARGEE_GET_CHARGE_LIMIT,
-    ENTITY_CHARGEE_LOCATION_SENSOR,
-    ENTITY_CHARGEE_SET_CHARGE_LIMIT,
-    ENTITY_CHARGEE_SOC_SENSOR,
-    ENTITY_CHARGEE_UPDATE_HA_BUTTON,
-    ENTITY_CHARGEE_WAKE_UP_BUTTON,
     ENTITY_CHARGER_CHARGING_SENSOR,
     ENTITY_CHARGER_GET_CHARGE_CURRENT,
     ENTITY_CHARGER_ON_OFF_SWITCH,
     ENTITY_CHARGER_PLUGGED_IN_SENSOR,
     ENTITY_CHARGER_SET_CHARGE_CURRENT,
-    NUMBER_CHARGEE_MAX_CHARGE_LIMIT,
-    NUMBER_CHARGEE_MIN_CHARGE_LIMIT,
+    ENTITY_DEVICE_GET_CHARGE_LIMIT,
+    ENTITY_DEVICE_LOCATION_SENSOR,
+    ENTITY_DEVICE_SET_CHARGE_LIMIT,
+    ENTITY_DEVICE_SOC_SENSOR,
+    ENTITY_DEVICE_UPDATE_HA_BUTTON,
+    ENTITY_DEVICE_WAKE_UP_BUTTON,
     NUMBER_CHARGER_MAX_CURRENT,
-    OPTION_CHARGEE_LOCATION_STATE_LIST,
+    NUMBER_DEVICE_MAX_CHARGE_LIMIT,
+    NUMBER_DEVICE_MIN_CHARGE_LIMIT,
     OPTION_CHARGER_CHARGING_STATE_LIST,
     OPTION_CHARGER_CONNECT_STATE_LIST,
+    OPTION_DEVICE_LOCATION_STATE_LIST,
 )
 from ..models.model_config import ConfigValueDict
 from .chargeable import Chargeable
@@ -91,7 +91,7 @@ class ChargerChargeableBase(HaDevice, ScOptionState, Charger, Chargeable):
         """Wake up chargeable device."""
 
         await self.async_option_press_entity_button(
-            ENTITY_CHARGEE_WAKE_UP_BUTTON, val_dict=val_dict
+            ENTITY_DEVICE_WAKE_UP_BUTTON, val_dict=val_dict
         )
 
     # ----------------------------------------------------------------------------
@@ -99,7 +99,7 @@ class ChargerChargeableBase(HaDevice, ScOptionState, Charger, Chargeable):
         """Force chargeable device to update data in HA."""
 
         await self.async_option_press_entity_button(
-            ENTITY_CHARGEE_UPDATE_HA_BUTTON, val_dict=val_dict
+            ENTITY_DEVICE_UPDATE_HA_BUTTON, val_dict=val_dict
         )
 
     # ----------------------------------------------------------------------------
@@ -109,10 +109,10 @@ class ChargerChargeableBase(HaDevice, ScOptionState, Charger, Chargeable):
 
         # 'device_tracker.tesla23m3_location_tracker' = 'not_home' or 'home'
         state = self.option_get_entity_string(
-            ENTITY_CHARGEE_LOCATION_SENSOR, val_dict=val_dict
+            ENTITY_DEVICE_LOCATION_SENSOR, val_dict=val_dict
         )
         state_list = self.option_get_list(
-            OPTION_CHARGEE_LOCATION_STATE_LIST, val_dict=val_dict
+            OPTION_DEVICE_LOCATION_STATE_LIST, val_dict=val_dict
         )
         if state is not None and state_list is not None:
             is_at_location = state in state_list
@@ -126,7 +126,7 @@ class ChargerChargeableBase(HaDevice, ScOptionState, Charger, Chargeable):
         """Get state of charge (SoC) of chargeable device."""
 
         return self.option_get_entity_integer(
-            ENTITY_CHARGEE_SOC_SENSOR, val_dict=val_dict
+            ENTITY_DEVICE_SOC_SENSOR, val_dict=val_dict
         )
 
     # ----------------------------------------------------------------------------
@@ -134,7 +134,7 @@ class ChargerChargeableBase(HaDevice, ScOptionState, Charger, Chargeable):
         """Get chargeable device charge limit."""
 
         return self.option_get_entity_number(
-            ENTITY_CHARGEE_GET_CHARGE_LIMIT, val_dict=val_dict
+            ENTITY_DEVICE_GET_CHARGE_LIMIT, val_dict=val_dict
         )
 
     # ----------------------------------------------------------------------------
@@ -144,17 +144,17 @@ class ChargerChargeableBase(HaDevice, ScOptionState, Charger, Chargeable):
         """Set chargeable device charge limit."""
 
         min_limit = self.option_get_entity_number_or_abort(
-            NUMBER_CHARGEE_MIN_CHARGE_LIMIT, val_dict=val_dict
+            NUMBER_DEVICE_MIN_CHARGE_LIMIT, val_dict=val_dict
         )
         max_limit = self.option_get_entity_number_or_abort(
-            NUMBER_CHARGEE_MAX_CHARGE_LIMIT, val_dict=val_dict
+            NUMBER_DEVICE_MAX_CHARGE_LIMIT, val_dict=val_dict
         )
         if not min_limit <= charge_limit <= max_limit:
             msg = f"Invalid charge limit {charge_limit}. Must be between {min_limit} and {max_limit} %%."
             raise ValueError(msg)
 
         await self.async_option_set_entity_number(
-            ENTITY_CHARGEE_SET_CHARGE_LIMIT, charge_limit, val_dict=val_dict
+            ENTITY_DEVICE_SET_CHARGE_LIMIT, charge_limit, val_dict=val_dict
         )
 
     # ----------------------------------------------------------------------------