+++ custom_components/solarcharger/config/config_options_flow.py
@@ -20,17 +20,17 @@ from homeassistant.helpers.selector import SelectSelector, SelectSelectorConfig
 
 from ..const import (
     DOMAIN,
-    ENTITY_CHARGEE_GET_CHARGE_LIMIT,
-    ENTITY_CHARGEE_LOCATION_SENSOR,
-    ENTITY_CHARGEE_SET_CHARGE_LIMIT,
-    ENTITY_CHARGEE_SOC_SENSOR,
-    ENTITY_CHARGEE_UPDATE_HA_BUTTON,
-    ENTITY_CHARGEE_WAKE_UP_BUTTON,
     ENTITY_CHARGER_CHARGING_SENSOR,
     ENTITY_CHARGER_GET_CHARGE_CURRENT,
     ENTITY_CHARGER_ON_OFF_SWITCH,
     ENTITY_CHARGER_PLUGGED_IN_SENSOR,
     ENTITY_CHARGER_SET_CHARGE_CURRENT,
+    ENTITY_DEVICE_GET_CHARGE_LIMIT,
+    ENTITY_DEVICE_LOCATION_SENSOR,
+    ENTITY_DEVICE_SET_CHARGE_LIMIT,
+    ENTITY_DEVICE_SOC_SENSOR,
+    ENTITY_DEVICE_UPDATE_HA_BUTTON,
+    ENTITY_DEVICE_WAKE_UP_BUTTON,
     ERROR_EMPTY_CHARGER_LIST,
     ERROR_NUMBER_FORMAT,
     ERROR_SUBENTRY_ID_NOT_FOUND,
@@ -42,8 +42,6 @@ from ..const import (
     NUMBER_CHARGE_LIMIT_THURSDAY,
     NUMBER_CHARGE_LIMIT_TUESDAY,
     NUMBER_CHARGE_LIMIT_WEDNESDAY,
-    NUMBER_CHARGEE_MAX_CHARGE_LIMIT,
-    NUMBER_CHARGEE_MIN_CHARGE_LIMIT,
     NUMBER_CHARGER_EFFECTIVE_VOLTAGE,
     NUMBER_CHARGER_MAX_CURRENT,
     NUMBER_CHARGER_MAX_SPEED,
@@ -61,20 +59,22 @@ from ..const import (
     NUMBER_DEFAULT_CHARGE_LIMIT_THURSDAY,
     NUMBER_DEFAULT_CHARGE_LIMIT_TUESDAY,
     NUMBER_DEFAULT_CHARGE_LIMIT_WEDNESDAY,
+    NUMBER_DEVICE_MAX_CHARGE_LIMIT,
+    NUMBER_DEVICE_MIN_CHARGE_LIMIT,
     NUMBER_POWER_MONITOR_DURATION,
     NUMBER_SUNRISE_ELEVATION_START_TRIGGER,
     NUMBER_SUNSET_ELEVATION_END_TRIGGER,
-    NUMBER_WAIT_CHARGEE_LIMIT_CHANGE,
-    NUMBER_WAIT_CHARGEE_UPDATE_HA,
-    NUMBER_WAIT_CHARGEE_WAKEUP,
     NUMBER_WAIT_CHARGER_AMP_CHANGE,
     NUMBER_WAIT_CHARGER_OFF,
     NUMBER_WAIT_CHARGER_ON,
-    OPTION_CHARGEE_LOCATION_STATE_LIST,
+    NUMBER_WAIT_DEVICE_LIMIT_CHANGE,
+    NUMBER_WAIT_DEVICE_UPDATE_HA,
+    NUMBER_WAIT_DEVICE_WAKEUP,
     OPTION_CHARGER_CHARGING_STATE_LIST,
     OPTION_CHARGER_CONNECT_STATE_LIST,
     OPTION_CHARGER_CONNECT_TRIGGER_LIST,
     OPTION_CHARGER_NAME,
+    OPTION_DEVICE_LOCATION_STATE_LIST,
     OPTION_GLOBAL_DEFAULT_ENTITIES,
     OPTION_GLOBAL_DEFAULTS_ID,
     OPTION_ID,
@@ -216,10 +216,10 @@ class ConfigOptionsFlowHandler(OptionsFlow):
             #####################################
             # Max/min charge limits
             self._optional(
-                subentry, NUMBER_CHARGEE_MIN_CHARGE_LIMIT, use_default
+                subentry, NUMBER_DEVICE_MIN_CHARGE_LIMIT, use_default
             ): NUMBER_ENTITY_SELECTOR,
             self._optional(
-                subentry, NUMBER_CHARGEE_MAX_CHARGE_LIMIT, use_default
+                subentry, NUMBER_DEVICE_MAX_CHARGE_LIMIT, use_default
             ): NUMBER_ENTITY_SELECTOR,
             # Charge limit defaults
             self._optional(
@@ -304,13 +304,13 @@ class ConfigOptionsFlowHandler(OptionsFlow):
             # Wait times
             #####################################
             self._optional(
-                subentry, NUMBER_WAIT_CHARGEE_WAKEUP, use_default
+                subentry, NUMBER_WAIT_DEVICE_WAKEUP, use_default
             ): NUMBER_ENTITY_SELECTOR,
             self._optional(
-                subentry, NUMBER_WAIT_CHARGEE_UPDATE_HA, use_default
+                subentry, NUMBER_WAIT_DEVICE_UPDATE_HA, use_default
             ): NUMBER_ENTITY_SELECTOR,
             self._optional(
-                subentry, NUMBER_WAIT_CHARGEE_LIMIT_CHANGE, use_default
+                subentry, NUMBER_WAIT_DEVICE_LIMIT_CHANGE, use_default
             ): NUMBER_ENTITY_SELECTOR,
             self._optional(
                 subentry, NUMBER_WAIT_CHARGER_ON, use_default
@@ -520,61 +520,61 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_ENTITY_SELECTOR,
             ),
             self._optional(
-                subentry, ENTITY_CHARGEE_SOC_SENSOR, use_default
+                subentry, ENTITY_DEVICE_SOC_SENSOR, use_default
             ): choose_selector(
                 api_entities,
-                ENTITY_CHARGEE_SOC_SENSOR,
+                ENTITY_DEVICE_SOC_SENSOR,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
             ),
             self._optional(
-                subentry, ENTITY_CHARGEE_GET_CHARGE_LIMIT, use_default
+                subentry, ENTITY_DEVICE_GET_CHARGE_LIMIT, use_default
             ): choose_selector(
                 api_entities,
-                ENTITY_CHARGEE_GET_CHARGE_LIMIT,
-                NUMBER_ENTITY_SELECTOR_READ_ONLY,
-                NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
-            ),
-            self._optional(
-                subentry, ENTITY_CHARGEE_SET_CHARGE_LIMIT, use_default
-            ): choose_selector(
-                api_entities,
-                ENTITY_CHARGEE_SET_CHARGE_LIMIT,
+                ENTITY_DEVICE_GET_CHARGE_LIMIT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
                 modifiable_if_solarcharger_entity=True,
             ),
             self._optional(
-                subentry, ENTITY_CHARGEE_LOCATION_SENSOR, use_default
+                subentry, ENTITY_DEVICE_SET_CHARGE_LIMIT, use_default
             ): choose_selector(
                 api_entities,
-                ENTITY_CHARGEE_LOCATION_SENSOR,
+                ENTITY_DEVICE_SET_CHARGE_LIMIT,
+                NUMBER_ENTITY_SELECTOR_READ_ONLY,
+                NUMBER_ENTITY_SELECTOR,
+                modifiable_if_solarcharger_entity=True,
+            ),
+            self._optional(
+                subentry, ENTITY_DEVICE_LOCATION_SENSOR, use_default
+            ): choose_selector(
+                api_entities,
+                ENTITY_DEVICE_LOCATION_SENSOR,
                 LOCATION_ENTITY_SELECTOR_READ_ONLY,
                 LOCATION_ENTITY_SELECTOR,
             ),
             self._optional(
-                subentry, OPTION_CHARGEE_LOCATION_STATE_LIST, use_default
+                subentry, OPTION_DEVICE_LOCATION_STATE_LIST, use_default
             ): choose_selector(
                 api_entities,
-                OPTION_CHARGEE_LOCATION_STATE_LIST,
+                OPTION_DEVICE_LOCATION_STATE_LIST,
                 TEXT_SELECTOR_READ_ONLY,
                 TEXT_SELECTOR,
             ),
             self._optional(
-                subentry, ENTITY_CHARGEE_WAKE_UP_BUTTON, use_default
+                subentry, ENTITY_DEVICE_WAKE_UP_BUTTON, use_default
             ): choose_selector(
                 api_entities,
-                ENTITY_CHARGEE_WAKE_UP_BUTTON,
+                ENTITY_DEVICE_WAKE_UP_BUTTON,
                 BUTTON_ENTITY_SELECTOR_READ_ONLY,
                 BUTTON_ENTITY_SELECTOR,
             ),
             # Turning on force HA update switch will override the in-built update HA button.
             self._optional(
-                subentry, ENTITY_CHARGEE_UPDATE_HA_BUTTON, use_default
+                subentry, ENTITY_DEVICE_UPDATE_HA_BUTTON, use_default
             ): choose_selector(
                 api_entities,
-                ENTITY_CHARGEE_UPDATE_HA_BUTTON,
+                ENTITY_DEVICE_UPDATE_HA_BUTTON,
                 BUTTON_ENTITY_SELECTOR_READ_ONLY,
                 BUTTON_ENTITY_SELECTOR,
             ),