+++ custom_components/solarcharger/number.py
@@ -33,9 +33,6 @@ from .const import (
     NUMBER_CHARGE_LIMIT_THURSDAY,
     NUMBER_CHARGE_LIMIT_TUESDAY,
     NUMBER_CHARGE_LIMIT_WEDNESDAY,
-    NUMBER_CHARGEE_CHARGE_LIMIT,
-    NUMBER_CHARGEE_MAX_CHARGE_LIMIT,
-    NUMBER_CHARGEE_MIN_CHARGE_LIMIT,
     NUMBER_CHARGER_EFFECTIVE_VOLTAGE,
     NUMBER_CHARGER_MAX_CURRENT,
     NUMBER_CHARGER_MAX_SPEED,
@@ -53,17 +50,20 @@ from .const import (
     NUMBER_DEFAULT_CHARGE_LIMIT_THURSDAY,
     NUMBER_DEFAULT_CHARGE_LIMIT_TUESDAY,
     NUMBER_DEFAULT_CHARGE_LIMIT_WEDNESDAY,
+    NUMBER_DEVICE_CHARGE_LIMIT,
+    NUMBER_DEVICE_MAX_CHARGE_LIMIT,
+    NUMBER_DEVICE_MIN_CHARGE_LIMIT,
     NUMBER_OCPP_PROFILE_ID,
     NUMBER_OCPP_PROFILE_STACK_LEVEL,
     NUMBER_POWER_MONITOR_DURATION,
     NUMBER_SUNRISE_ELEVATION_START_TRIGGER,
     NUMBER_SUNSET_ELEVATION_END_TRIGGER,
-    NUMBER_WAIT_CHARGEE_LIMIT_CHANGE,
-    NUMBER_WAIT_CHARGEE_UPDATE_HA,
-    NUMBER_WAIT_CHARGEE_WAKEUP,
     NUMBER_WAIT_CHARGER_AMP_CHANGE,
     NUMBER_WAIT_CHARGER_OFF,
     NUMBER_WAIT_CHARGER_ON,
+    NUMBER_WAIT_DEVICE_LIMIT_CHANGE,
+    NUMBER_WAIT_DEVICE_UPDATE_HA,
+    NUMBER_WAIT_DEVICE_WAKEUP,
 )
 from .entity import (
     SolarChargerEntity,
@@ -195,13 +195,13 @@ CONFIG_NUMBER_LIST: tuple[
     #####################################
     # Used as local device entity for OCPP only. Others come with own entity.
     (
-        NUMBER_CHARGEE_CHARGE_LIMIT,
+        NUMBER_DEVICE_CHARGE_LIMIT,
         [
             SolarChargerEntityType.TYPE_LOCAL_OCPP,
             SolarChargerEntityType.TYPE_LOCAL_USER_CUSTOM,
         ],
         NumberEntityDescription(
-            key=NUMBER_CHARGEE_CHARGE_LIMIT,
+            key=NUMBER_DEVICE_CHARGE_LIMIT,
             entity_category=EntityCategory.CONFIG,
             native_unit_of_measurement=PERCENTAGE,
             native_min_value=0,
@@ -372,10 +372,10 @@ CONFIG_NUMBER_LIST: tuple[
         ),
     ),
     (
-        NUMBER_WAIT_CHARGEE_WAKEUP,
+        NUMBER_WAIT_DEVICE_WAKEUP,
         SolarChargerEntityType.TYPE_LOCALHIDDEN_GLOBAL,
         NumberEntityDescription(
-            key=NUMBER_WAIT_CHARGEE_WAKEUP,
+            key=NUMBER_WAIT_DEVICE_WAKEUP,
             entity_category=EntityCategory.CONFIG,
             device_class=NumberDeviceClass.DURATION,
             native_unit_of_measurement=UnitOfTime.SECONDS,
@@ -385,10 +385,10 @@ CONFIG_NUMBER_LIST: tuple[
         ),
     ),
     (
-        NUMBER_WAIT_CHARGEE_UPDATE_HA,
+        NUMBER_WAIT_DEVICE_UPDATE_HA,
         SolarChargerEntityType.TYPE_LOCALHIDDEN_GLOBAL,
         NumberEntityDescription(
-            key=NUMBER_WAIT_CHARGEE_UPDATE_HA,
+            key=NUMBER_WAIT_DEVICE_UPDATE_HA,
             entity_category=EntityCategory.CONFIG,
             device_class=NumberDeviceClass.DURATION,
             native_unit_of_measurement=UnitOfTime.SECONDS,
@@ -398,10 +398,10 @@ CONFIG_NUMBER_LIST: tuple[
         ),
     ),
     (
-        NUMBER_WAIT_CHARGEE_LIMIT_CHANGE,
+        NUMBER_WAIT_DEVICE_LIMIT_CHANGE,
         SolarChargerEntityType.TYPE_LOCALHIDDEN_GLOBAL,
         NumberEntityDescription(
-            key=NUMBER_WAIT_CHARGEE_LIMIT_CHANGE,
+            key=NUMBER_WAIT_DEVICE_LIMIT_CHANGE,
             entity_category=EntityCategory.CONFIG,
             device_class=NumberDeviceClass.DURATION,
             native_unit_of_measurement=UnitOfTime.SECONDS,
@@ -470,10 +470,10 @@ CONFIG_NUMBER_LIST: tuple[
     #####################################
     # Max/min charge limit config.
     (
-        NUMBER_CHARGEE_MIN_CHARGE_LIMIT,
+        NUMBER_DEVICE_MIN_CHARGE_LIMIT,
         SolarChargerEntityType.TYPE_LOCALHIDDEN_GLOBAL,
         NumberEntityDescription(
-            key=NUMBER_CHARGEE_MIN_CHARGE_LIMIT,
+            key=NUMBER_DEVICE_MIN_CHARGE_LIMIT,
             entity_category=EntityCategory.CONFIG,
             native_unit_of_measurement=PERCENTAGE,
             native_min_value=0,
@@ -482,10 +482,10 @@ CONFIG_NUMBER_LIST: tuple[
         ),
     ),
     (
-        NUMBER_CHARGEE_MAX_CHARGE_LIMIT,
+        NUMBER_DEVICE_MAX_CHARGE_LIMIT,
         SolarChargerEntityType.TYPE_LOCALHIDDEN_GLOBAL,
         NumberEntityDescription(
-            key=NUMBER_CHARGEE_MAX_CHARGE_LIMIT,
+            key=NUMBER_DEVICE_MAX_CHARGE_LIMIT,
             entity_category=EntityCategory.CONFIG,
             native_unit_of_measurement=PERCENTAGE,
             native_min_value=0,